Refactor into a scroller manager
